Make test server customizable
