+++ test/server/BUILD
@@ -23,6 +23,7 @@ java_library(
     name = "grakn-setup",
     srcs = [
         "GraknSetup.java",
+        "GraknCoreSetup.java",
     ],
     deps = [
         "@maven//:commons_io_commons_io",
